More fix

+++ assignment4/src/cli.h
@@ -8,7 +8,7 @@
 
 typedef struct {
     size_t n_iter;
-    float diff_c;
+    double diff_c;
     int verbose;
 } CliArgs;
